refactor joins into module
