+++ app/src/main/java/fho/kdvs/favorite/broadcast/FavoriteBroadcastRepository.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import fho.kdvs.global.BaseRepository
 import fho.kdvs.global.database.FavoriteBroadcastDao
 import fho.kdvs.global.database.FavoriteBroadcastEntity
-import fho.kdvs.global.database.ShowBroadcastFavoriteJoin
+import fho.kdvs.global.database.joins.ShowBroadcastFavoriteJoin
 import fho.kdvs.global.extensions.toLiveData
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
