+++ app/src/main/java/fho/kdvs/global/database/BroadcastDao.kt
@@ -3,6 +3,7 @@ package fho.kdvs.global.database
 import androidx.lifecycle.LiveData
 import androidx.room.*
 import androidx.room.OnConflictStrategy.REPLACE
+import fho.kdvs.global.database.joins.ShowTimeslotBroadcastJoin
 import io.reactivex.Flowable
 import org.threeten.bp.LocalDate
