+++ app/src/main/java/fho/kdvs/global/database/FavoriteBroadcastDao.kt
@@ -6,6 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy.REPLACE
 import androidx.room.Query
 import androidx.room.Transaction
+import fho.kdvs.global.database.joins.ShowBroadcastFavoriteJoin
 import io.reactivex.Flowable
 
 @Dao
