+++ app/src/main/java/fho/kdvs/global/database/FavoriteTrackDao.kt
@@ -6,6 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy.REPLACE
 import androidx.room.Query
 import androidx.room.Transaction
+import fho.kdvs.global.database.joins.ShowBroadcastTrackFavoriteJoin
 import io.reactivex.Flowable
 
 @Dao
