+++ app/src/main/java/fho/kdvs/global/database/joins/BroadcastTrackJoin.kt
@@ -1,0 +1,14 @@
+package fho.kdvs.global.database.joins
+
+import androidx.room.Embedded
+import androidx.room.Relation
+import fho.kdvs.global.database.BroadcastEntity
+import fho.kdvs.global.database.TrackEntity
+
+class BroadcastTrackJoin {
+    @Embedded
+    var broadcast: BroadcastEntity? = null
+
+    @Relation(parentColumn = "broadcastId", entityColumn = "broadcastId")
+    var track: List<TrackEntity> = ArrayList()
+}
