+++ app/src/main/java/fho/kdvs/global/database/joins/Join.kt
@@ -1,0 +1,3 @@
+package fho.kdvs.global.database.joins
+
+open class Join
