+++ app/src/main/java/fho/kdvs/global/database/joins/ShowBroadcastJoin.kt
@@ -1,0 +1,14 @@
+package fho.kdvs.global.database.joins
+
+import androidx.room.Embedded
+import androidx.room.Relation
+import fho.kdvs.global.database.BroadcastEntity
+import fho.kdvs.global.database.ShowEntity
+
+class ShowBroadcastJoin {
+    @Embedded
+    var show: ShowEntity? = null
+
+    @Relation(parentColumn = "id", entityColumn = "showId")
+    var broadcast: List<BroadcastEntity> = ArrayList()
+}
